+++ CHANGELOG.md
@@ -17,6 +17,7 @@ IMPROVEMENTS
 * Changes to tool names to be more consistent. See [#121](https://github.com/hashicorp/terraform-mcp-server/pull/121)
 * Implement pagination utility. See [#121](https://github.com/hashicorp/terraform-mcp-server/pull/121)
 * Updating `mark3labs/mcp-go` and `hashicorp/tfe-go` versions. See [#121](https://github.com/hashicorp/terraform-mcp-server/pull/121)
+* Implementing TLS for the http mode of the MCP server. See [#168](https://github.com/hashicorp/terraform-mcp-server/pull/168)
 * Implemented rate limiting with the MCP server. See [#155](https://github.com/hashicorp/terraform-mcp-server/pull/155)
 * Improved and expanded README with usage instructions for Claude Code. See [#173](https://github.com/hashicorp/terraform-mcp-server/pull/173)
 * Improved and expanded README with usage instructions for Cursor. See [#173](https://github.com/hashicorp/terraform-mcp-server/pull/173)